project/requirements.txt
torch
torchvision
python-socketio
numpy
